Refactor main.tsx and address TS/ESLint warnings
The code in main.tsx has been reformatted for improved legibility. The refactor includes handling TypeScript and ESLint warnings by altering the previous "ts-ignore" directive to a "ts-expect-error" one. Changes in syntax have been made including the addition of a semi-colon and minor changes in spaces for cleaner and more quality code.